Backfill a test for the clone bug
Refs https://github.com/dependents/node-dependency-tree/pull/58
